Consolidate LongBench v2 evaluation into single unified file
- Merge longbench_v2_*.py files into single longbench_v2.py
- Support multiple backends: direct, vLLM, official scripts
- Automatic backend selection based on environment
- Auto-detect native chat templates from tokenizer_config.json
- Unified configuration with longbench_mode parameter
- Clean up redundant files and simplify architecture
- Update documentation with new unified approach

Benefits:
- Single file to maintain instead of 3 separate ones
- Automatic backend selection for best performance
- Support for all evaluation scenarios (development, production, official)
- Cleaner codebase with unified API
